% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{chilean_potential}
\alias{chilean_potential}
\title{The potential foodweb of the intertidal zone in central chile}
\format{
a network
}
\usage{
chilean_potential
}
\description{
A trophic network with 107 species present in the  intertidal zone of central Chile.
The food web was reconstructed from the Kefi et al. 2015
}
\keyword{datasets}
