% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{chilean_weighted}
\alias{chilean_weighted}
\title{The weighted foodweb of the intertidal zone in central chile}
\format{
a network
}
\usage{
chilean_weighted
}
\description{
A trophic network with 107 species present in the  intertidal zone of central Chile.
The food web was reconstructed from the Kefi et al. 2015
}
\references{
Kefi, Sonia, Eric L. Berlow, Evie A. Wieters, Lucas N. Joppa, Spencer A. Wood, Ulrich Brose, and Sergio A. Navarrete. "Network structure beyond food webs: mapping non trophic and trophic interactions on Chilean rocky shores." Ecology 96, no. 1 (2015.
}
\keyword{datasets}
