% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Powerlaw.R
\name{degree_distribution}
\alias{degree_distribution}
\title{Degree distribution of the network}
\usage{
degree_distribution(Network, name)
}
\arguments{
\item{Network}{a trophic network of class network}

\item{name}{a categorical variable that represent
the distribution model}
}
\value{
exports three principal results:
1. A list with network degree distribution results
2. A list with the fit values and AIC of the distribution models
3. A Ghraph of the degree distribution with the models adjust
In the first results, k represent the degree of the network and cumulative
the probability that each specie could be have this degree (pk).
}
\description{
Calculate the degree distribution of the network. Then, adjust
the exponential, power law and power law truncated distribution
to the network degree distribution and calculate de AIC to  evaluate
the best model. Finally, do a plot of the degree distribution in a log-log scale
and adjust the three models mentioned above
}
\author{
Derek Corcoran <derek.corcoran.barrios@gmail.com>

Isidora Avila <msavila@uc.cl>
}
