% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{dist}
\alias{dist}
\title{A toymodel distance matrix}
\format{
a distance matrix
}
\usage{
dist
}
\description{
A distance matrix used for demonstration of rewiring capabilities
}
\keyword{datasets}
