% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Less_Connected}
\alias{Less_Connected}
\title{A sparsely connected foodweb}
\format{
a network
}
\usage{
Less_Connected
}
\description{
A network with 30 species and 47 interactions.
This network has a connectance of 0.03
}
\seealso{
\code{\link{More_Connected}}
}
\keyword{datasets}
