% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{More_Connected}
\alias{More_Connected}
\title{A densely connected foodweb}
\format{
a network
}
\usage{
More_Connected
}
\description{
A trophic network with 30 species and 222 trophic interactions.
This foodweb has a connectance of 0.3
}
\seealso{
\code{\link{Less_Connected}}
}
\keyword{datasets}
