% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extintions.R
\name{Mostconnected}
\alias{Mostconnected}
\title{Extnction analyses from more to less conected}
\usage{
Mostconnected(Network)
}
\arguments{
\item{Network}{a trophic network of class network}
}
\value{
exports data frame with the characteristics of the network after every
extintion
}
\description{
Generates an AMPL dat file from a Stack in which each file is the projection
of a species distribution model into a time slice
}
\examples{
data("net")
Mostconnected(Network = net)
}
\author{
Derek Corcoran <derek.corcoran.barrios@gmail.com>

Javier Fajardo <javierfajnolla@gmail.com >
}
