% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extintions.R
\name{MostconnectedExp}
\alias{MostconnectedExp}
\title{Extnction analyses from more to less conected in a network}
\usage{
MostconnectedExp(Network)
}
\arguments{
\item{Network}{a trophic network of class network}
}
\value{
exports data frame with the characteristics of the network after every
extintion. The resulting data frame contains seven columns
}
\description{
It takes a network and it calculates wich species is the most conected
of the network, then it extinguishes that species, and calculates the
secundary extintions. After that it recalculates the conections and
repeats the process until every species becomes extinct.
}
\examples{
data("net")
MostconnectedExp(Network = net)
}
\seealso{
[NetworkExtintion::ExtinctionOrder()]
}
\author{
Derek Corcoran <derek.corcoran.barrios@gmail.com>

Isidora Avila <msavila@uc.cl>
}
