% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{net}
\alias{net}
\title{A toymodel trophic network}
\format{
a network
}
\usage{
net
}
\description{
A trophic network with 10 species where the first four species are primery producters
}
\keyword{datasets}
