% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extintions.R
\name{RandomExtinctions}
\alias{RandomExtinctions}
\title{Random extinction}
\usage{
RandomExtinctions(Network, nsim = 10)
}
\arguments{
\item{Network}{a trophic network of class network}

\item{nsim}{number of simulations}
}
\value{
exports data frame with the characteristics of the network after every
extintion, and a graph with the mean and 95% interval
}
\description{
Generates a null model by generating random extinction histories and calculating
the mean and standard deviation of the accumulated secondary extinctions developed
by making n random extinction histories
}
\examples{
#first example
data("net")
ExtinctionOrder(Network = net, Order = c(1,2,3,4,5,6,7,8,9,10))
#Second example
data("net")
ExtinctionOrder(Network = net, Order = c(2,8,9))

}
\author{
Derek Corcoran <derek.corcoran.barrios@gmail.com>

M.Isidora Avila Thieme <msavila@uc.cl>
}
